+++ client/cody-ui/src/Chat.tsx
@@ -133,8 +133,8 @@ export const Chat: React.FunctionComponent<ChatProps> = ({
     needsEmailVerificationNotice: NeedsEmailVerificationNotice,
     contextStatusComponent: ContextStatusComponent,
     contextStatusComponentProps = {},
-    abortMessageInProgressComponent,
-    onAbortMessageInProgress,
+    abortMessageInProgressComponent: AbortMessageInProgressButton,
+    onAbortMessageInProgress = () => {},
     isCodyEnabled,
 }) => {
     const [inputRows, setInputRows] = useState(5)
@@ -266,8 +266,6 @@ export const Chat: React.FunctionComponent<ChatProps> = ({
                     copyButtonOnSubmit={copyButtonOnSubmit}
                     submitButtonComponent={SubmitButton}
                     chatInputClassName={chatInputClassName}
-                    abortMessageInProgressComponent={abortMessageInProgressComponent}
-                    onAbortMessageInProgress={onAbortMessageInProgress}
                 />
             )}
 
@@ -285,6 +283,11 @@ export const Chat: React.FunctionComponent<ChatProps> = ({
                         )}
                     </div>
                 ) : null}
+                {messageInProgress && AbortMessageInProgressButton && (
+                    <div className={classNames(styles.abortButtonContainer)}>
+                        <AbortMessageInProgressButton onAbortMessageInProgress={onAbortMessageInProgress} />
+                    </div>
+                )}
                 <div className={styles.textAreaContainer}>
                     <TextArea
                         className={classNames(styles.chatInput, chatInputClassName)}